rename new plugins
