+++ src/runtime_src/core/edge/ps_kernels/profiling/aie2_profile_config/profile_event_configuration.h
@@ -25,8 +25,8 @@
 #include "xaiefal/xaiefal.hpp"
 #include "xaiengine.h"
 #include "xdp/profile/database/static_info/aie_constructs.h"
-#include "xdp/profile/plugin/aie_profile_new/x86/aie_profile_kernel_config.h"
-#include "xdp/profile/plugin/aie_profile_new/aie_profile_defs.h"
+#include "xdp/profile/plugin/aie_profile/x86/aie_profile_kernel_config.h"
+#include "xdp/profile/plugin/aie_profile/aie_profile_defs.h"
 
 // This struct encapsulates all of the internal configuration information
 // for a single AIE tile
