+++ src/runtime_src/core/edge/ps_kernels/profiling/aie2_trace_config/aie2_trace_config.cpp
@@ -16,13 +16,14 @@
 #include <cstring>
 #include <vector>
 
+#include "event_configuration.h"
+#include "xaiengine.h"
+
 #include "core/edge/include/sk_types.h"
 #include "core/edge/user/shim.h"
-#include "event_configuration.h"
 #include "xaiefal/xaiefal.hpp"
-#include "xaiengine.h"
 #include "xdp/profile/database/static_info/aie_constructs.h"
-#include "xdp/profile/plugin/aie_trace_new/x86/aie_trace_kernel_config.h"
+#include "xdp/profile/plugin/aie_trace/x86/aie_trace_kernel_config.h"
 
 // User private data structure container (context object) definition
 class xrtHandles : public pscontext
