+++ src/runtime_src/core/edge/ps_kernels/profiling/aie2_trace_config/event_configuration.h
@@ -23,7 +23,7 @@
 #include <vector>
 #include "xaiefal/xaiefal.hpp"
 #include "xaiengine.h"
-#include "xdp/profile/plugin/aie_trace_new/x86/aie_trace_kernel_config.h"
+#include "xdp/profile/plugin/aie_trace/x86/aie_trace_kernel_config.h"
 
 constexpr uint32_t ES1_TRACE_COUNTER = 1020;
 constexpr uint32_t ES2_TRACE_COUNTER = 0x3FF00;
