+++ src/runtime_src/xdp/profile/plugin/aie_profile/edge/aie_profile.cpp
@@ -21,18 +21,18 @@
 #include <memory>
 #include <cstring>
 
-#include "../aie_profile_defs.h"
 #include "aie_profile.h"
-#include "core/edge/user/shim.h"
+
 #include "core/common/message.h"
 #include "core/common/time.h"
 #include "core/common/xrt_profiling.h"
+#include "core/edge/user/shim.h"
 #include "core/include/xrt/xrt_kernel.h"
-
 #include "xdp/profile/database/database.h"
 #include "xdp/profile/database/static_info/aie_constructs.h"
 #include "xdp/profile/database/static_info/pl_constructs.h"
-#include "xdp/profile/plugin/aie_profile_new/aie_profile_metadata.h"
+#include "xdp/profile/plugin/aie_profile/aie_profile_defs.h"
+#include "xdp/profile/plugin/aie_profile/aie_profile_metadata.h"
 
 namespace {
   static void* fetchAieDevInst(void* devHandle)
