+++ src/runtime_src/xdp/profile/plugin/aie_profile/x86/aie_profile.cpp
@@ -22,16 +22,16 @@
 #include <cstring>
 
 #include "aie_profile.h"
+
 #include "core/common/message.h"
 #include "core/common/time.h"
 #include "core/common/xrt_profiling.h"
 #include "core/include/xrt/xrt_kernel.h"
-
 #include "xdp/profile/database/database.h"
 #include "xdp/profile/database/static_info/aie_constructs.h"
 #include "xdp/profile/database/static_info/pl_constructs.h"
-#include "xdp/profile/plugin/aie_profile_new/aie_profile_metadata.h"
-#include "xdp/profile/plugin/aie_profile_new/x86/aie_profile_kernel_config.h"
+#include "xdp/profile/plugin/aie_profile/aie_profile_metadata.h"
+#include "xdp/profile/plugin/aie_profile/x86/aie_profile_kernel_config.h"
 
 
 constexpr uint32_t ALIGNMENT_SIZE = 4096;
