+++ src/runtime_src/xdp/profile/plugin/aie_trace/CMakeLists.txt
