+++ src/runtime_src/xdp/profile/plugin/aie_trace/x86/aie_trace.h
@@ -20,7 +20,7 @@
 #include <cstdint>
 
 #include "core/include/xrt/xrt_kernel.h"
-#include "xdp/profile/plugin/aie_trace_new/aie_trace_impl.h"
+#include "xdp/profile/plugin/aie_trace/aie_trace_impl.h"
 
 namespace xdp {
